split Indexed and Monad
